Fix compression method 1
Previously, the computed diff frame would use 0 to represent a pixel that
was unchanged between the 2 frames. The problem is that 0 is also a valid
pixel value. Only store 0 or 1 in the diff frame and reference the current
frame for current pixel values.

Also, adjust the logic of when to close a run since it's possible for a run
to be closing at the max of 128 pixels AND be flipping between run types.